Added 1 endpoint to retrieve checkLists : /getCheckLists
API Endpoint is a get endpoint and expect one parameter in the URL(studentNumber) eg)/getCheckLists?studentNumber=1234567, This retrieves all checkLists with their items that are associated with all courses that a student with studentNumber is enrolled in. The JSON returned has a success propertt and the following properties are JSON Objects which have in order of appearance the attributes : checkListID,topic,courseName,courseID,visibility,deadline,checkListItemID,checkListItem. These can all be used for logic and grouping checklist items together using courses and/or checklists